Replace tokens only when they are at least 1/2 expired
This way we save CPU resources on the server by not creating new
on every request.  Instead we validate the current token and we
keep it in place until it is at least 1/2 expired.  Since any of
the XSRF tokens are likely to have a 12-24 hour time window this
means we will only produce a new token once or twice a day per
user, cutting our per-request costs.

Signed-off-by: Shawn O. Pearce <sop@google.com>