Push null handling for JsonSerializer.printJson to the caller
When outputting a JSON object we try to avoid putting a field
into the object dictionary if the field's value is null.  As
such the caller of any serializer has already done the "is it
null" test and handled the result of that decision, so we are
only wasting CPU time by putting it into the serializer's own
printJson method too.

However this means we must handle null at the parameter level
for any proxy methods.  Which is pretty easy to do since its
only a small if test around the argument.

Signed-off-by: Shawn O. Pearce <sop@google.com>