Declare CallbackHandle for JSON based callback generation
A callback handle allows the application to install a global
JavaScript function in the browser window and arrange for that
function to deserialize a JSON graph into a normal Java object,
to be passed to an AsyncCallback's onSuccess method.

This is useful for services which may be making cross-domain calls,
or which need to generate a callback into the application after
showing an HTML page inside of an iframe.

Signed-off-by: Shawn O. Pearce <sop@google.com>