Slightly randomize the callback function numbers to avoid attacks
By using the time of day that the callback handle code was first
initialized in the browser we can reduce the risk that someone
can guess our callback function names, especially for a sign in
sort of callback.  This should help to prevent external sites from
calling our callbacks during a JSONP return, unless we gave them
the name of the function up front.

Signed-off-by: Shawn O. Pearce <sop@google.com>