Define AllowCrossSiteRequest annotation to disable the XSRF token
Some JSON requests might want to permit cross-site requests, such
as if they are reading information from the server that isn't tied
to a specific individual, and/or is public information.  These can
be annotated with the AllowCrossSiteRequest annotation, which skips
the token in the HTTP requests and skips validation in the server.

Signed-off-by: Shawn O. Pearce <sop@google.com>