Cleanup error handling and add GET and JSON-in-script callbacks
The JSON-in-script format honors an approach similar to what Yahoo!
permits in their JSON-in-script APIs, which is (roughly) any valid
JavaScript identifier, so function names like "foo.bar" can be used
by a client application.

The GET style requests support invoking methods with positional
arguments in the query string, rather than in the POST body, and
is thus suitable for use in a <script> style tag, especially once
combined with the callback parameter.

Error handling has been cleaned up to expose nothing to the client,
unless the method implementation itself calls onFailure().  Those
will be reported to the application by sending the message text.

Signed-off-by: Shawn O. Pearce <sop@google.com>