Define HostPageCache annotation to supply initial RPC results
The new HostPageCache annotation permits the server to include any
JavaScript objects into the host page and lets the GWT module read
them back through an RPC style interface.  If the object isn't in
the page, or the object has already been "consumed" once then the
call will use a traditional RPC to obtain the result.

Signed-off-by: Shawn O. Pearce <sop@google.com>