Use xsrfKey in request payload rather than HTTP headers
Some browers such as WebKit on Android won't send (or maybe it is
receive) arbitrary HTTP headers.  Instead of using a custom HTTP
header we now embed the XSRF token into the payload of either a
request or response.

This change forced a refactoring of the request generation code
resulting in the smaller output files due to better common string
pooling.

Signed-off-by: Shawn O. Pearce <sop@google.com>