Improve ActiveCall.noCache to better handle common no-cache cases
Forcing the date to something long in the past ensures proxies
won't cache the content.  Its also faster to set a header to
a string than to format it out from a timestamp.

We also include must-revalidate in Cache-Control to ensure a proxy
doesn't return stale content while waiting for an update from
the remote server.

Signed-off-by: Shawn O. Pearce <sop@google.com>