Work around GWT 1.5.3 bug parsing "2009-01-08" as a java.sql.Date
GWT 1.5.3 and earlier all parse "2009-01-08" (or time "20:09:15")
by splitting on the field separator and then parsing as integers
through Integer.decode().  The fields that start with "0", e.g.
"08" in the date example above, are parsed as octal, and "8" is
not a valid octal digit, so the parse fails and the client gets
an IllegalArgumentException thrown back.

This bug only manifests itself in compiled mode, where the emul
layer is being used to supply java.sql.Timestamp.  In hosted
mode the JRE is supplying a non-buggy java.sql.Timestamp.  This
made the issue that much more difficult to track down.

GWT's trunk still contains this problem, so it may be possible
for GWT 1.6 to ship with this bug.  For the foreseeable future
we'll need to carry our own java.sql.Date, java.sql.Time and
java.sql.Timestamp parsers.

Signed-off-by: Shawn O. Pearce <sop@google.com>