Fix the java.sql.Timestamp serialization to always use UTC
I really meant for the format of a java.sql.Timestamp to be
an ISO datetime format, but to always be in UTC so that the
client and server don't need to worry about timezones, and
we don't have to encode it in the payload.

Unfortunately the stock implementations of Timestamp's parse and
format methods always ssume local timezone, resulting in some
level of confusion as values were converted back and forth from
the database to the UI.

We now use our own formatting routines which take care to ensure
only UTC is used.

Signed-off-by: Shawn O. Pearce <sop@google.com>