Tighten the JSON servlet so GET requests cannot ever allow XSRF methods
If a method requires XSRF protection it cannot be served over a GET
request, even if the caller found a way to populate our ActiveCall
structure with the valid token data.  Evil sites can make a request
using <SCRIPT SRC="..us.."> and override the Array constructor in
the source page to gain access to any array members we may have in
the resulting stream.

There are currently no known exploits where the browser would be
able to parse and evaluate the object block we always return back,
but that doesn't mean one won't be discovered in the future.  By
blocking any of these calls here we can help reduce our exposure
if an attach was discovered.

Signed-off-by: Shawn O. Pearce <sop@google.com>