Fix server side parsing of UTF-8 encoded JSON payloads
The server was treating a UTF-8 encoded payload as though it was
encoded in the server's default character set, resulting in data
with non-ASCII characters being parsed incorrectly.

The client now always sends "; charset=utf-8" in its requests
and we require that the character set on the request always
be specified as UTF-8.  We also parse the payload ourselves
as UTF-8 data, to avoid any problems with the container doing
something magical (and incorrect) with the character encoding.

Signed-off-by: Shawn O. Pearce <sop@google.com>