Allow user defined generic types to be parameterized and serialized
GWT provides us the concrete type information when processing a user
defined class that has type parameters declared on it.  Unfortunately
we can't see the type parameters, so we can't determine how the inputs
need to be passed down through the serializer chain.  Instead by using
a unique serializer class name per permutation of type parameters we
are able to just inline the type information all the way down.

Signed-off-by: Shawn O. Pearce <sop@google.com>