SignedToken: Use URL-safe encoding
There is a problem that if the token is used as a HTTP Request Parameter
in URL, the verification of `SignedToken` will be always a failure.

	BASE64 scope : A-Z a-z 0-9 + / =
	URL transfer scope: + SPACE / ? % # & =

For an example, when the token contains char '+', the parameter of token
will treat this char as a SPACE character.  So, it's strongly needed
for `new` and `check` to support `isUrlSafe` flag to avoid the runtime
verification problems.

This affects Gerrit's email confirmation API. If the token contains '+'
the email verification will fail.

Bug: Issue 12424
Change-Id: Id71926f1231a57df0da7eecae6033c0d28f36eee