SignedToken supports `isUrlSafe` when BASE64 works
There is a problem that if the token is used as an HTTP Request Parameter
in URL, the verification of `SignedToken` will be always a failure.

	BASE64 scope : A-Z a-z 0-9 + / =
	URL transfer scope: + SPACE / ? % # & =

For example, When the token contains char '+', the parameter of the
token will treat this char as a SPACE character.  So, it's strongly
needed for `new` and `check` to support `isUrlSafe` flag to avoid the
runtime verification problems.

P.S.: Problem was found in Gerrit email confirm function API, if token
contains '+' the verify will be failed.

Bug: Issue 12424
Change-Id: Id71926f1231a57df0da7eecae6033c0d28f36eee