Update patch set 4
Patch Set 4:

David:

I want to upload a new snapshot verison of gwtjsonrpc to https://oss.sonatype.org/ for a further test integrate with gerrit.


But it seems like I have no privillege of `deploy snapshot` with it.

Can you give me an access right of it or other ways to solve it .

Patch-set: 4